Tweak recommender comparison criteria and display
Instead of picking 15 anime with scores above average (5 from the top, 5 from
the middle, 5 from the bottom scores) for each user, only pick 5 from the top
and 2 from the middle, and choose anime with a normalized score 0.6 above the
mean or higher.

This should provide a better comparison under the assumption that it is more
important to return items with a high rating than it is to cover the whole
lange of items above average.