Change the sync script to use user's last_change rather than last_update
Using last_update is silly because it means that all users are synced, either
in one direction or the other, because last_update is guaranteed to be
different even if anime data is the same.

After the changes from last commits, last_change is a good indicator of an
actual data change.