Tweak user sample seleciton again
Now each bucket is limited to a maximum of 20 clusters, and the conditions for
inserting an user change depending on the index of the cluster in the sequence
(stricter for earlier clusters); users that do not fit any clusters are simply
ignored.

This gives an overall better coverage of anime, and better user diversification.

Other changes:

- change buckets to fixed size (15 or 45)
- use only completed anime count for bucketing, not
  the sum of completed and dropped anime