Fix franchise aggregation
The way it was performed was fatally flawed: it was throwing away some of the
relations when loading from the database (because it was building a map without
computing the transitive closure) and then it was losing some more information
by performing the map update after the iteration.

Now it constructs the map in a single iteration after loading the anime list,
and exploiting mutability.

Fixes #27